calendar_back/pyproject.toml
[tool.poetry]
name = "calendar-back"
version = "0.1.0"
description = ""
authors = ["byungyong <quddyddlrk@naver.com>"]
readme = "README.md"
packages = [{include = "calendar_back"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
django-cors-headers = "^4.2.0"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
drf-yasg = "^1.21.7"
drf-spectacular = "^0.26.4"
mysqlclient = "^2.2.0"
gunicorn = "^21.2.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
